fix api test with removing test that doesn't have @Test
