--- app/src/test/java/com/danteyu/studio/artbooktesting/data/source/api/ApiAbstract.kt
@@ -1,79 +1,0 @@
-/*
- * Copyright 2021 Dante Yu
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.danteyu.studio.artbooktesting.data.source.api
-
-import com.squareup.moshi.Moshi
-import com.squareup.moshi.kotlin.reflect.KotlinJsonAdapterFactory
-import okhttp3.mockwebserver.MockResponse
-import okhttp3.mockwebserver.MockWebServer
-import okio.buffer
-import okio.source
-import org.junit.After
-import org.junit.Before
-import org.junit.runner.RunWith
-import org.junit.runners.JUnit4
-import retrofit2.Retrofit
-import retrofit2.converter.moshi.MoshiConverterFactory
-import java.io.IOException
-import java.nio.charset.StandardCharsets
-
-/**
- * Created by George Yu in Jun. 2021.
- */
-@RunWith(JUnit4::class)
-open class ApiAbstract<T> {
-
-    lateinit var mockWebServer: MockWebServer
-
-    @Throws(IOException::class)
-    @Before
-    fun setup() {
-        mockWebServer = MockWebServer()
-        mockWebServer.start()
-    }
-
-    @Throws(IOException::class)
-    @After
-    fun tearDown() {
-        mockWebServer.shutdown()
-    }
-
-    @Throws(IOException::class)
-    fun enqueueResponse(fileName: String) {
-        enqueueResponse(fileName, emptyMap())
-    }
-
-    @Throws(IOException::class)
-    private fun enqueueResponse(fileName: String, headers: Map<String, String> = emptyMap()) {
-        val inputStream = javaClass.classLoader!!.getResourceAsStream("api-response/$fileName")
-        val source = inputStream.source().buffer()
-        val mockResponse = MockResponse()
-        for ((key, value) in headers) {
-            mockResponse.addHeader(key, value)
-        }
-        mockWebServer.enqueue(mockResponse.setBody(source.readString(StandardCharsets.UTF_8)))
-    }
-
-    fun createService(clazz: Class<T>): T {
-        val moshi = Moshi.Builder().add(KotlinJsonAdapterFactory()).build()
-
-        return Retrofit.Builder()
-            .baseUrl(mockWebServer.url("/"))
-            .addConverterFactory(MoshiConverterFactory.create(moshi))
-            .build()
-            .create(clazz)
-    }
-}
